G925F OI9 Code

+++ sound/soc/samsung/seiren/seiren-dma.c
@@ -20,13 +20,14 @@
 #include <linux/kthread.h>
 #include <linux/pm_runtime.h>
 #include <linux/iommu.h>
-#include <linux/dma/dma-pl330.h>
 
 #include <sound/soc.h>
 #include <sound/pcm_params.h>
 #include <sound/exynos.h>
 
 #include <asm/dma.h>
+#include <mach/hardware.h>
+#include <mach/dma.h>
 #include <mach/map.h>
 
 #include "seiren.h"
@@ -77,7 +78,7 @@ static int of_dma_match_channel(struct device_node *np, const char *name,
 	return 0;
 }
 
-static unsigned long esa_dma_request(enum dma_ch dma_ch,
+static unsigned esa_dma_request(enum dma_ch dma_ch,
 				struct samsung_dma_req *param,
 				struct device *dev, char *ch_name)
 {
@@ -129,7 +130,7 @@ static unsigned long esa_dma_request(enum dma_ch dma_ch,
 					__func__, rtd->ch, rtd->peri);
 			writel(rtd->peri, rtd->regs + DMA_PERI);
 
-			return (unsigned long)ch;
+			return (unsigned)ch;
 		}
 	}
 	esa_err("%s: dmas property of %s missing\n", __func__, ch_name);
@@ -143,7 +144,7 @@ err:
 	return -EIO;
 }
 
-static int esa_dma_release(unsigned  long ch, void *param)
+static int esa_dma_release(unsigned ch, void *param)
 {
 	struct runtime *rtd = sdi.rtd[ch];
 
@@ -155,7 +156,7 @@ static int esa_dma_release(unsigned  long ch, void *param)
 	return 0;
 }
 
-static int esa_dma_config(unsigned long ch, struct samsung_dma_config *param)
+static int esa_dma_config(unsigned ch, struct samsung_dma_config *param)
 {
 	struct runtime *rtd = sdi.rtd[ch];
 	void __iomem *regs = rtd->regs;
@@ -178,7 +179,7 @@ static int esa_dma_config(unsigned long ch, struct samsung_dma_config *param)
 	return 0;
 }
 
-static int esa_dma_prepare(unsigned long ch, struct samsung_dma_prep *param)
+static int esa_dma_prepare(unsigned ch, struct samsung_dma_prep *param)
 {
 	struct runtime *rtd = sdi.rtd[ch];
 	void __iomem *regs = rtd->regs;
@@ -201,7 +202,7 @@ static int esa_dma_prepare(unsigned long ch, struct samsung_dma_prep *param)
 	return 0;
 }
 
-static int esa_dma_trigger(unsigned long ch)
+static int esa_dma_trigger(unsigned ch)
 {
 	struct runtime *rtd = sdi.rtd[ch];
 
@@ -210,7 +211,7 @@ static int esa_dma_trigger(unsigned long ch)
 	return 0;
 }
 
-static int esa_dma_getposition(unsigned long ch, dma_addr_t *src, dma_addr_t *dst)
+static int esa_dma_getposition(unsigned ch, dma_addr_t *src, dma_addr_t *dst)
 {
 	struct runtime *rtd = sdi.rtd[ch];
 
@@ -220,7 +221,7 @@ static int esa_dma_getposition(unsigned long ch, dma_addr_t *src, dma_addr_t *ds
 	return 0;
 }
 
-static int esa_dma_flush(unsigned long ch)
+static int esa_dma_flush(unsigned ch)
 {
 	struct runtime *rtd = sdi.rtd[ch];
